pyproject.toml
[project]
name = "streamlit-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "ollama>=0.4.7",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
]
